Renamed reducer and action SamplesGrid.js to sampleGrid.js
