+++ mxcube3/ui/actions/sampleGrid.js
