+++ mxcube3/ui/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import login from './login';
 import queue from './queue';
 import queueGUI from './queueGUI';
-import sampleGrid from './SamplesGrid';
+import sampleGrid from './sampleGrid';
 import sampleChanger from './sampleChanger';
 import taskForm from './taskForm';
 import sampleview from './sampleview';
