+++ mxcube3/ui/reducers/sampleGrid.js
